+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Auto-generate and use Pydantic models for the API requests payloads #62
 * Use server-side multi-circuits jobs API #63
 * Add job completion progress bar #63
+* Allow overriding any backend option in `AQTResource.run` #64
+* Only return raw memory data when the `memory` option is set #64
 * Implement the `ProviderV1` interface for `AQTProvider` #65
 * Set User-Agent with package and platform information for HTTP requests #65